+++ src/modbam2bed.c
@@ -33,7 +33,7 @@ void *pileup_worker(void *arg) {
         files, j.chr, j.start, j.end,
         j.args.read_group, j.args.tag_name, j.args.tag_value,
         j.args.threshold, j.args.mod_base, j.args.combine,
-        j.args.hts_maxcnt);
+        j.args.hts_maxcnt, j.args.min_mapQ);
     destroy_filesets(files);
     free(arg);
     return pileup;
@@ -58,7 +58,7 @@ void process_region(arguments_t args, const char *chr, int start, int end, char 
         args.bam, chr, start, end,
         args.read_group, args.tag_name, args.tag_value,
         args.threshold, args.mod_base, args.combine,
-        args.hts_maxcnt);
+        args.hts_maxcnt, args.min_mapQ);
     if (pileup == NULL) return;
 
     init_output_buffers(bed_files);