+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.9.1]
+### Added
+- `--map_q` command line option to filter reads by minimum mapping quality.
+### Fixed
+- The default mapping quality was erroneously 1 not 0.
+
 ## [v0.9.0]
 ### Added
 - "Other" modified base column to extended output. For example, when using